refactor: split stateful zone implementation
